+++ tests/actions/test_core_actions.py
@@ -1,12 +1,10 @@
 from __future__ import annotations
 
-import datetime as dt
-import json
 import logging
 from contextlib import contextmanager
 from datetime import datetime, timezone
 from types import ModuleType
-from typing import TYPE_CHECKING, Iterator, Type
+from typing import TYPE_CHECKING, Iterator
 from unittest import mock
 
 import pytest
@@ -24,7 +22,6 @@ from great_expectations.checkpoint.actions import (
     SlackNotificationAction,
     SNSNotificationAction,
     UpdateDataDocsAction,
-    ValidationAction,
 )
 from great_expectations.checkpoint.checkpoint import Checkpoint, CheckpointResult
 from great_expectations.core.batch import IDDict, LegacyBatchDefinition
@@ -56,198 +53,116 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
+SUITE_A: str = "SUITE_A"
+SUITE_B: str = "SUITE_B"
+BATCH_ID_A: str = "my_datasource-my_first_asset"
+BATCH_ID_B: str = "my_datasource-my_second_asset"
+
+
 @pytest.fixture
-def mock_context(mocker: MockerFixture):
-    context = mocker.MagicMock(spec=AbstractDataContext)
-    set_context(context)
-    return context
+def mocked_posthog(mocker: MockerFixture):
+    yield mocker.patch("posthog.capture")
 
 
-class MockTeamsResponse:
-    def __init__(self, status_code: int, raise_for_status: bool = False):
-        self.status_code = status_code
-        self._raise_for_status = raise_for_status
-        self.text = "test_text"
-
-    def raise_for_status(self):
-        if self._raise_for_status:
-            raise requests.exceptions.HTTPError("test")
-
-
-class MockSlackResponse:
-    def __init__(self, status_code):
-        self.status_code = status_code
-        self.text = "ok"
-        self.content = json.dumps({"ok": "True"})
-
-    def json(self):
-        return {"ok": "True"}
-
-    def raise_for_status(self):
-        pass
-
-
-class MockCloudResponse:
-    def __init__(self, status_code):
-        self.status_code = status_code
-        self.text = "test_text"
-        self.content = json.dumps({"ok": "True"})
-
-
-@pytest.mark.unit
-def test_api_action_create_payload(mock_context):
-    mock_validation_results = []
-    expected_payload = {
-        "test_suite_name": "my_suite",
-        "data_asset_name": "my_schema.my_table",
-        "validation_results": [],
-    }
-    api_notification_action = APINotificationAction(
-        name="my_api_notification", url="http://www.example.com"
+@pytest.fixture
+def checkpoint_result(mocker: MockerFixture):
+    utc_datetime = datetime.fromisoformat("2024-04-01T20:51:18.077262").replace(tzinfo=timezone.utc)
+    return CheckpointResult(
+        run_id=RunIdentifier(run_time=utc_datetime),
+        run_results={
+            ValidationResultIdentifier(
+                expectation_suite_identifier=ExpectationSuiteIdentifier(
+                    name=SUITE_A,
+                ),
+                run_id=RunIdentifier(run_name="prod_20240401"),
+                batch_identifier=BATCH_ID_A,
+            ): ExpectationSuiteValidationResult(
+                success=True,
+                statistics={"successful_expectations": 3, "evaluated_expectations": 3},
+                results=[],
+                suite_name=SUITE_A,
+            ),
+            ValidationResultIdentifier(
+                expectation_suite_identifier=ExpectationSuiteIdentifier(
+                    name=SUITE_B,
+                ),
+                run_id=RunIdentifier(run_name="prod_20240402"),
+                batch_identifier=BATCH_ID_B,
+            ): ExpectationSuiteValidationResult(
+                success=True,
+                statistics={"successful_expectations": 2, "evaluated_expectations": 2},
+                results=[],
+                suite_name=SUITE_B,
+            ),
+        },
+        checkpoint_config=Checkpoint(
+            name="test-checkpoint",
+            validation_definitions=[
+                mocker.MagicMock(spec=ValidationDefinition),
+                mocker.MagicMock(spec=ValidationDefinition),
+            ],
+        ),
     )
-    payload = api_notification_action.create_payload(
-        "my_schema.my_table", "my_suite", mock_validation_results
+
+
+@pytest.fixture
+def checkpoint_result_with_assets(mocker: MockerFixture):
+    SUITE_A: str = "SUITE_A"
+    SUITE_B: str = "SUITE_B"
+    BATCH_ID_A: str = "my_datasource-my_first_asset"
+    BATCH_ID_B: str = "my_datasource-my_second_asset"
+    utc_datetime = datetime.fromisoformat("2024-04-01T20:51:18.077262").replace(tzinfo=timezone.utc)
+    return CheckpointResult(
+        run_id=RunIdentifier(run_time=utc_datetime),
+        run_results={
+            ValidationResultIdentifier(
+                expectation_suite_identifier=ExpectationSuiteIdentifier(
+                    name=SUITE_A,
+                ),
+                run_id=RunIdentifier(run_name="prod_20240401"),
+                batch_identifier=BATCH_ID_A,
+            ): ExpectationSuiteValidationResult(
+                success=True,
+                statistics={"successful_expectations": 3, "evaluated_expectations": 3},
+                results=[],
+                suite_name=SUITE_A,
+                meta={
+                    "active_batch_definition": LegacyBatchDefinition(
+                        datasource_name="test_environment",
+                        data_connector_name="general_azure_data_connector",
+                        data_asset_name="asset_1",
+                        batch_identifiers=IDDict(
+                            {"name": "alex", "timestamp": "20200809", "price": "1000"}
+                        ),
+                    )
+                },
+                result_url="www.testing",
+            ),
+            ValidationResultIdentifier(
+                expectation_suite_identifier=ExpectationSuiteIdentifier(
+                    name=SUITE_B,
+                ),
+                run_id=RunIdentifier(run_name="prod_20240402"),
+                batch_identifier=BATCH_ID_B,
+            ): ExpectationSuiteValidationResult(
+                success=True,
+                statistics={"successful_expectations": 2, "evaluated_expectations": 2},
+                results=[],
+                suite_name=SUITE_B,
+                meta={
+                    "active_batch_definition": LegacyBatchDefinition(
+                        datasource_name="test_environment",
+                        data_connector_name="general_azure_data_connector",
+                        data_asset_name="asset_2_two_wow_whoa_vroom",
+                        batch_identifiers=IDDict(
+                            {"name": "alex", "timestamp": "20200809", "price": "1000"}
+                        ),
+                    ),
+                },
+            ),
+        },
+        checkpoint_config=mocker.Mock(spec=Checkpoint, name="my_checkpoint"),
     )
-    assert payload == expected_payload
-
-
-class TestActionSerialization:
-    EXAMPLE_SLACK_WEBHOOK = "https://hooks.slack.com/services/test/slack/webhook"
-    EXAMPLE_TEAMS_WEBHOOK = "https://hooks.microsoft.com/services/test/teams/webhook"
-    EXAMPLE_API_KEY = "testapikey"
-    EXAMPLE_SMTP_ADDRESS = "smtp.test.com"
-    EXAMPLE_SMTP_PORT = 587
-    EXAMPLE_EMAILS = "bob@gmail.com, jim@hotmail.com"
-    EXAMPLE_SITE_NAMES = ["one_site", "two_site", "red_site", "blue_site"]
-    EXAMPLE_SNS_TOPIC_ARN = "my_test_arn"
-    EXAMPLE_URL = "http://www.example.com"
-
-    ACTION_INIT_PARAMS = {
-        SlackNotificationAction: {
-            "name": "my_slack_action",
-            "slack_webhook": EXAMPLE_SLACK_WEBHOOK,
-        },
-        MicrosoftTeamsNotificationAction: {
-            "name": "my_teams_action",
-            "teams_webhook": EXAMPLE_TEAMS_WEBHOOK,
-        },
-        OpsgenieAlertAction: {"name": "my_opsgenie_action", "api_key": EXAMPLE_API_KEY},
-        EmailAction: {
-            "name": "my_email_action",
-            "smtp_address": EXAMPLE_SMTP_ADDRESS,
-            "smtp_port": EXAMPLE_SMTP_PORT,
-            "receiver_emails": EXAMPLE_EMAILS,
-        },
-        UpdateDataDocsAction: {"name": "my_data_docs_action", "site_names": EXAMPLE_SITE_NAMES},
-        SNSNotificationAction: {"name": "my_sns_action", "sns_topic_arn": EXAMPLE_SNS_TOPIC_ARN},
-        APINotificationAction: {"name": "my_api_action", "url": EXAMPLE_URL},
-    }
-
-    SERIALIZED_ACTIONS = {
-        SlackNotificationAction: {
-            "name": "my_slack_action",
-            "notify_on": "all",
-            "notify_with": None,
-            "renderer": {
-                "class_name": "SlackRenderer",
-                "module_name": "great_expectations.render.renderer.slack_renderer",
-            },
-            "show_failed_expectations": False,
-            "slack_channel": None,
-            "slack_token": None,
-            "slack_webhook": EXAMPLE_SLACK_WEBHOOK,
-            "type": "slack",
-        },
-        MicrosoftTeamsNotificationAction: {
-            "name": "my_teams_action",
-            "notify_on": "all",
-            "renderer": {
-                "class_name": "MicrosoftTeamsRenderer",
-                "module_name": "great_expectations.render.renderer.microsoft_teams_renderer",
-            },
-            "teams_webhook": EXAMPLE_TEAMS_WEBHOOK,
-            "type": "microsoft",
-        },
-        OpsgenieAlertAction: {
-            "name": "my_opsgenie_action",
-            "api_key": EXAMPLE_API_KEY,
-            "notify_on": "failure",
-            "priority": "P3",
-            "region": None,
-            "renderer": {
-                "class_name": "OpsgenieRenderer",
-                "module_name": "great_expectations.render.renderer.opsgenie_renderer",
-            },
-            "tags": None,
-            "type": "opsgenie",
-        },
-        EmailAction: {
-            "name": "my_email_action",
-            "notify_on": "all",
-            "notify_with": None,
-            "receiver_emails": EXAMPLE_EMAILS,
-            "renderer": {
-                "class_name": "EmailRenderer",
-                "module_name": "great_expectations.render.renderer.email_renderer",
-            },
-            "sender_alias": None,
-            "sender_login": None,
-            "sender_password": None,
-            "smtp_address": EXAMPLE_SMTP_ADDRESS,
-            "smtp_port": str(EXAMPLE_SMTP_PORT),
-            "type": "email",
-            "use_ssl": None,
-            "use_tls": None,
-        },
-        UpdateDataDocsAction: {
-            "name": "my_data_docs_action",
-            "site_names": EXAMPLE_SITE_NAMES,
-            "type": "update_data_docs",
-        },
-        SNSNotificationAction: {
-            "name": "my_sns_action",
-            "sns_message_subject": None,
-            "sns_topic_arn": EXAMPLE_SNS_TOPIC_ARN,
-            "type": "sns",
-        },
-        APINotificationAction: {
-            "name": "my_api_action",
-            "type": "api",
-            "url": EXAMPLE_URL,
-        },
-    }
-
-    @pytest.mark.parametrize(
-        "action_class, init_params",
-        [(k, v) for k, v in ACTION_INIT_PARAMS.items()],
-        ids=[k.__name__ for k in ACTION_INIT_PARAMS],
-    )
-    @pytest.mark.unit
-    def test_action_serialization_and_deserialization(
-        self,
-        mock_context,
-        action_class: Type[ValidationAction],
-        init_params: dict,
-    ):
-        expected = self.SERIALIZED_ACTIONS[action_class]
-
-        action = action_class(**init_params)
-        json_dict = action.json()
-        actual = json.loads(json_dict)
-
-        assert actual == expected
-
-    @pytest.mark.parametrize(
-        "action_class, serialized_action",
-        [(k, v) for k, v in SERIALIZED_ACTIONS.items()],
-        ids=[k.__name__ for k in SERIALIZED_ACTIONS],
-    )
-    @pytest.mark.unit
-    def test_action_deserialization(
-        self, action_class: Type[ValidationAction], serialized_action: dict
-    ):
-        actual = action_class.parse_obj(serialized_action)
-        assert isinstance(actual, action_class)
 
 
 @contextmanager
@@ -262,111 +177,25 @@ def mock_not_imported_module(
         setattr(parent_module, target_name, original)
 
 
-class TestV1ActionRun:
-    suite_a: str = "suite_a"
-    suite_b: str = "suite_b"
-    batch_id_a: str = "my_datasource-my_first_asset"
-    batch_id_b: str = "my_datasource-my_second_asset"
-
-    @pytest.fixture
-    def mocked_posthog(self, mocker: MockerFixture):
-        yield mocker.patch("posthog.capture")
-
-    @pytest.fixture
-    def checkpoint_result(self, mocker: MockerFixture):
-        utc_datetime = datetime.fromisoformat("2024-04-01T20:51:18.077262").replace(
-            tzinfo=timezone.utc
+class TestAPINotificationAction:
+    @pytest.mark.unit
+    def test_create_payload(self, mock_context):
+        mock_validation_results = []
+        expected_payload = {
+            "test_suite_name": "my_suite",
+            "data_asset_name": "my_schema.my_table",
+            "validation_results": [],
+        }
+        api_notification_action = APINotificationAction(
+            name="my_api_notification", url="http://www.example.com"
         )
-        return CheckpointResult(
-            run_id=RunIdentifier(run_time=utc_datetime),
-            run_results={
-                ValidationResultIdentifier(
-                    expectation_suite_identifier=ExpectationSuiteIdentifier(
-                        name=self.suite_a,
-                    ),
-                    run_id=RunIdentifier(run_name="prod_20240401"),
-                    batch_identifier=self.batch_id_a,
-                ): ExpectationSuiteValidationResult(
-                    success=True,
-                    statistics={"successful_expectations": 3, "evaluated_expectations": 3},
-                    results=[],
-                    suite_name=self.suite_a,
-                ),
-                ValidationResultIdentifier(
-                    expectation_suite_identifier=ExpectationSuiteIdentifier(
-                        name=self.suite_b,
-                    ),
-                    run_id=RunIdentifier(run_name="prod_20240402"),
-                    batch_identifier=self.batch_id_b,
-                ): ExpectationSuiteValidationResult(
-                    success=True,
-                    statistics={"successful_expectations": 2, "evaluated_expectations": 2},
-                    results=[],
-                    suite_name=self.suite_b,
-                ),
-            },
-            checkpoint_config=mocker.Mock(spec=Checkpoint, name="my_checkpoint"),
+        payload = api_notification_action.create_payload(
+            "my_schema.my_table", "my_suite", mock_validation_results
         )
-
-    @pytest.fixture
-    def checkpoint_result_with_assets(self, mocker: MockerFixture):
-        utc_datetime = datetime.fromisoformat("2024-04-01T20:51:18.077262").replace(
-            tzinfo=timezone.utc
-        )
-        return CheckpointResult(
-            run_id=RunIdentifier(run_time=utc_datetime),
-            run_results={
-                ValidationResultIdentifier(
-                    expectation_suite_identifier=ExpectationSuiteIdentifier(
-                        name=self.suite_a,
-                    ),
-                    run_id=RunIdentifier(run_name="prod_20240401"),
-                    batch_identifier=self.batch_id_a,
-                ): ExpectationSuiteValidationResult(
-                    success=True,
-                    statistics={"successful_expectations": 3, "evaluated_expectations": 3},
-                    results=[],
-                    suite_name=self.suite_a,
-                    meta={
-                        "active_batch_definition": LegacyBatchDefinition(
-                            datasource_name="test_environment",
-                            data_connector_name="general_azure_data_connector",
-                            data_asset_name="asset_1",
-                            batch_identifiers=IDDict(
-                                {"name": "alex", "timestamp": "20200809", "price": "1000"}
-                            ),
-                        )
-                    },
-                    result_url="www.testing",
-                ),
-                ValidationResultIdentifier(
-                    expectation_suite_identifier=ExpectationSuiteIdentifier(
-                        name=self.suite_b,
-                    ),
-                    run_id=RunIdentifier(run_name="prod_20240402"),
-                    batch_identifier=self.batch_id_b,
-                ): ExpectationSuiteValidationResult(
-                    success=True,
-                    statistics={"successful_expectations": 2, "evaluated_expectations": 2},
-                    results=[],
-                    suite_name=self.suite_b,
-                    meta={
-                        "active_batch_definition": LegacyBatchDefinition(
-                            datasource_name="test_environment",
-                            data_connector_name="general_azure_data_connector",
-                            data_asset_name="asset_2_two_wow_whoa_vroom",
-                            batch_identifiers=IDDict(
-                                {"name": "alex", "timestamp": "20200809", "price": "1000"}
-                            ),
-                        ),
-                    },
-                ),
-            },
-            checkpoint_config=mocker.Mock(spec=Checkpoint, name="my_checkpoint"),
-        )
+        assert payload == expected_payload
 
     @pytest.mark.unit
-    def test_APINotificationAction_run(self, checkpoint_result: CheckpointResult, mocked_posthog):
+    def test_run(self, checkpoint_result: CheckpointResult, mocked_posthog):
         url = "http://www.example.com"
         action = APINotificationAction(name="my_action", url=url)
 
@@ -378,20 +207,22 @@ class TestV1ActionRun:
             headers={"Content-Type": "application/json"},
             data=[
                 {
-                    "data_asset_name": self.batch_id_a,
-                    "test_suite_name": self.suite_a,
+                    "data_asset_name": BATCH_ID_A,
+                    "test_suite_name": SUITE_A,
                     "validation_results": [],
                 },
                 {
-                    "data_asset_name": self.batch_id_b,
-                    "test_suite_name": self.suite_b,
+                    "data_asset_name": BATCH_ID_B,
+                    "test_suite_name": SUITE_B,
                     "validation_results": [],
                 },
             ],
         )
 
+
+class TestEmailAction:
     @pytest.mark.unit
-    def test_EmailAction_equality(self):
+    def test_equality(self):
         """I know, this one seems silly. But this was a bug."""
         a = EmailAction(
             name="my_action",
@@ -425,7 +256,7 @@ class TestV1ActionRun:
             ),
         ],
     )
-    def test_EmailAction_run(
+    def test_run(
         self,
         checkpoint_result: CheckpointResult,
         emails: str,
@@ -447,10 +278,10 @@ class TestV1ActionRun:
 
         # Should contain suite names and other relevant domain object identifiers in the body
         run_results = tuple(checkpoint_result.run_results.values())
-        suite_a = run_results[0].suite_name
-        suite_b = run_results[1].suite_name
+        SUITE_A = run_results[0].suite_name
+        SUITE_B = run_results[1].suite_name
         mock_html = mock_send_email.call_args.kwargs["html"]
-        assert suite_a in mock_html and suite_b in mock_html
+        assert SUITE_A in mock_html and SUITE_B in mock_html
 
         mock_send_email.assert_called_once_with(
             title=mock.ANY,
@@ -467,7 +298,7 @@ class TestV1ActionRun:
         assert out == {"email_result": mock_send_email()}
 
     @pytest.mark.unit
-    def test_EmailAction_run_smptp_address_substitution(
+    def test_run_smptp_address_substitution(
         self, checkpoint_result: CheckpointResult, mocked_posthog
     ):
         config_provider = project_manager.get_config_provider()
@@ -518,10 +349,10 @@ class TestV1ActionRun:
             use_tls=mock.ANY,
         )
 
+
+class TestMicrosoftTeamsNotificationAction:
     @pytest.mark.unit
-    def test_MicrosoftTeamsNotificationAction_run(
-        self, checkpoint_result: CheckpointResult, mocked_posthog
-    ):
+    def test_run(self, checkpoint_result: CheckpointResult, mocked_posthog):
         action = MicrosoftTeamsNotificationAction(name="my_action", teams_webhook="test")
 
         with mock.patch.object(Session, "post") as mock_post:
@@ -540,7 +371,7 @@ class TestV1ActionRun:
         assert body[1]["text"] == "Validation Result (1 of 2) ✅"
         assert body[2]["facts"] == [
             {"title": "Data Asset name: ", "value": "--"},
-            {"title": "Suite name: ", "value": "suite_a"},
+            {"title": "Suite name: ", "value": "SUITE_A"},
             {
                 "title": "Run name: ",
                 "value": "prod_20240401",
@@ -555,7 +386,7 @@ class TestV1ActionRun:
         assert body[3]["text"] == "Validation Result (2 of 2) ✅"
         assert body[4]["facts"] == [
             {"title": "Data Asset name: ", "value": "--"},
-            {"title": "Suite name: ", "value": "suite_b"},
+            {"title": "Suite name: ", "value": "SUITE_B"},
             {
                 "title": "Run name: ",
                 "value": "prod_20240402",
@@ -567,9 +398,7 @@ class TestV1ActionRun:
         ]
 
     @pytest.mark.unit
-    def test_MicrosoftTeamsNotificationAction_run_webhook_substitution(
-        self, checkpoint_result: CheckpointResult, mocked_posthog
-    ):
+    def test_run_webhook_substitution(self, checkpoint_result: CheckpointResult, mocked_posthog):
         config_provider = project_manager.get_config_provider()
         assert isinstance(config_provider, mock.Mock)  # noqa: TID251 # just using for the instance compare
 
@@ -596,6 +425,24 @@ class TestV1ActionRun:
             microsoft_teams_webhook=MS_TEAMS_WEBHOOK_VALUE,
         )
 
+    @pytest.mark.integration
+    def test_run_integration(
+        self,
+        checkpoint_result: CheckpointResult,
+    ):
+        # Necessary to retrieve config provider
+        gx.get_context(mode="ephemeral")
+
+        action = MicrosoftTeamsNotificationAction(
+            name="test-action", teams_webhook="${GX_MS_TEAMS_WEBHOOK}"
+        )
+        result = action.run(checkpoint_result=checkpoint_result)
+        assert result == {
+            "microsoft_teams_notification_result": "Microsoft Teams notification succeeded."
+        }
+
+
+class TestOpsgenieAlertAction:
     @pytest.mark.unit
     @pytest.mark.parametrize(
         "success, message",
@@ -604,9 +451,7 @@ class TestV1ActionRun:
             pytest.param(False, "failed!", id="failure"),
         ],
     )
-    def test_OpsgenieAlertAction_run(
-        self, checkpoint_result: CheckpointResult, success: bool, message: str
-    ):
+    def test_run(self, checkpoint_result: CheckpointResult, success: bool, message: str):
         action = OpsgenieAlertAction(name="my_action", api_key="test", notify_on="all")
         checkpoint_result.success = success
 
@@ -617,10 +462,10 @@ class TestV1ActionRun:
         assert message in mock_post.call_args.kwargs["json"]["message"]
         assert output == {"opsgenie_alert_result": True}
 
+
+class TestPagerdutyAlertAction:
     @pytest.mark.unit
-    def test_PagerdutyAlertAction_run_emits_events(
-        self, checkpoint_result: CheckpointResult, mocker: MockerFixture
-    ):
+    def test_run_emits_events(self, checkpoint_result: CheckpointResult, mocker: MockerFixture):
         from great_expectations.checkpoint import actions
 
         with mock_not_imported_module(actions, "pypd", mocker):
@@ -676,9 +521,7 @@ class TestV1ActionRun:
     )
     @mock.patch("pypd.EventV2.create")
     @pytest.mark.unit
-    def test_PagerdutyAlertAction_run_does_not_emit_events(
-        self, mock_pypd_event, checkpoint_result: CheckpointResult
-    ):
+    def test_run_does_not_emit_events(self, mock_pypd_event, checkpoint_result: CheckpointResult):
         action = PagerdutyAlertAction(
             name="my_action", api_key="test", routing_key="test", notify_on="failure"
         )
@@ -690,8 +533,10 @@ class TestV1ActionRun:
 
         mock_pypd_event.assert_not_called()
 
+
+class TestSlackNotificationAction:
     @pytest.mark.unit
-    def test_SlackNotificationAction_equality(self):
+    def test_equality(self):
         """I kow, this one seems silly. But this was a bug."""
         a = SlackNotificationAction(name="my_action", slack_webhook="test", notify_on="all")
         b = SlackNotificationAction(name="my_action", slack_webhook="test", notify_on="all")
@@ -699,7 +544,7 @@ class TestV1ActionRun:
         assert a == b
 
     @pytest.mark.unit
-    def test_SlackNotificationAction_run(self, checkpoint_result: CheckpointResult, mocked_posthog):
+    def test_run(self, checkpoint_result: CheckpointResult, mocked_posthog):
         action = SlackNotificationAction(name="my_action", slack_webhook="test", notify_on="all")
 
         with mock.patch.object(Session, "post") as mock_post:
@@ -719,14 +564,14 @@ class TestV1ActionRun:
                         "type": "section",
                         "text": {
                             "type": "mrkdwn",
-                            "text": "*Asset*: __no_data_asset_name__  *Expectation Suite*: suite_a",
+                            "text": "*Asset*: __no_data_asset_name__  *Expectation Suite*: SUITE_A",
                         },
                     },
                     {
                         "type": "section",
                         "text": {
                             "type": "mrkdwn",
-                            "text": "*Asset*: __no_data_asset_name__  *Expectation Suite*: suite_b",
+                            "text": "*Asset*: __no_data_asset_name__  *Expectation Suite*: SUITE_B",
                         },
                     },
                     {"type": "divider"},
@@ -737,9 +582,7 @@ class TestV1ActionRun:
         assert output == {"slack_notification_result": "Slack notification succeeded."}
 
     @pytest.mark.unit
-    def test_SlackNotificationAction_run_with_assets(
-        self, checkpoint_result_with_assets: CheckpointResult, mocked_posthog
-    ):
+    def test_run_with_assets(self, checkpoint_result_with_assets: CheckpointResult, mocked_posthog):
         action = SlackNotificationAction(name="my_action", slack_webhook="test", notify_on="all")
 
         with mock.patch.object(Session, "post") as mock_post:
@@ -759,7 +602,7 @@ class TestV1ActionRun:
                         "type": "section",
                         "text": {
                             "type": "mrkdwn",
-                            "text": "*Asset*: asset_1  *Expectation Suite*: suite_a  "
+                            "text": "*Asset*: asset_1  *Expectation Suite*: SUITE_A  "
                             "<www.testing?slack=true|View Results>",
                         },
                     },
@@ -768,7 +611,7 @@ class TestV1ActionRun:
                         "text": {
                             "type": "mrkdwn",
                             "text": "*Asset*: asset_2_two_wow_whoa_vroom  "
-                            "*Expectation Suite*: suite_b",
+                            "*Expectation Suite*: SUITE_B",
                         },
                     },
                     {"type": "divider"},
@@ -779,7 +622,7 @@ class TestV1ActionRun:
         assert output == {"slack_notification_result": "Slack notification succeeded."}
 
     @pytest.mark.unit
-    def test_SlackNotificationAction_grabs_data_docs_pages(
+    def test_grabs_data_docs_pages(
         self, checkpoint_result_with_assets: CheckpointResult, mocked_posthog
     ):
         action = SlackNotificationAction(name="my_action", slack_webhook="test", notify_on="all")
@@ -817,7 +660,7 @@ class TestV1ActionRun:
                         "type": "section",
                         "text": {
                             "type": "mrkdwn",
-                            "text": "*Asset*: asset_1  *Expectation Suite*: suite_a  "
+                            "text": "*Asset*: asset_1  *Expectation Suite*: SUITE_A  "
                             "<www.testing?slack=true|View Results>",
                         },
                     },
@@ -833,7 +676,7 @@ class TestV1ActionRun:
                         "text": {
                             "type": "mrkdwn",
                             "text": "*Asset*: asset_2_two_wow_whoa_vroom  "
-                            "*Expectation Suite*: suite_b",
+                            "*Expectation Suite*: SUITE_B",
                         },
                     },
                     {
@@ -858,9 +701,33 @@ class TestV1ActionRun:
         assert output == {"slack_notification_result": "Slack notification succeeded."}
 
     @pytest.mark.unit
-    def test_SNSNotificationAction_run(
-        self, sns, checkpoint_result: CheckpointResult, mocked_posthog
+    def test_variable_substitution_webhook(self, mock_context, checkpoint_result, mocked_posthog):
+        action = SlackNotificationAction(name="my_action", slack_webhook="${SLACK_WEBHOOK}")
+
+        with mock.patch("great_expectations.checkpoint.actions.send_slack_notification"):
+            action.run(checkpoint_result)
+
+        mock_context.config_provider.substitute_config.assert_called_once_with("${SLACK_WEBHOOK}")
+
+    @pytest.mark.unit
+    def test_variable_substitution_token_and_channel(
+        self, mock_context, checkpoint_result, mocked_posthog
     ):
+        action = SlackNotificationAction(
+            name="my_action", slack_token="${SLACK_TOKEN}", slack_channel="${SLACK_CHANNEL}"
+        )
+
+        with mock.patch("great_expectations.checkpoint.actions.send_slack_notification"):
+            action.run(checkpoint_result)
+
+        assert mock_context.config_provider.substitute_config.call_count == 2
+        mock_context.config_provider.substitute_config.assert_any_call("${SLACK_CHANNEL}")
+        mock_context.config_provider.substitute_config.assert_any_call("${SLACK_TOKEN}")
+
+
+class TestSNSNotificationAction:
+    @pytest.mark.unit
+    def test_run(self, sns, checkpoint_result: CheckpointResult, mocked_posthog):
         subj_topic = "test-subj"
         created_subj = sns.create_topic(Name=subj_topic)
         arn = created_subj.get("TopicArn")
@@ -873,8 +740,10 @@ class TestV1ActionRun:
         result = action.run(checkpoint_result=checkpoint_result)
         assert "Successfully posted results" in result["result"]
 
+
+class TestUpdateDataDocsAction:
     @pytest.mark.unit
-    def test_UpdateDataDocsAction_equality(self):
+    def test_equality(self):
         """I kow, this one seems silly. But this was a bug for other actions."""
         a = UpdateDataDocsAction(name="my_action")
         b = UpdateDataDocsAction(name="my_action")
@@ -882,9 +751,7 @@ class TestV1ActionRun:
         assert a == b
 
     @pytest.mark.unit
-    def test_UpdateDataDocsAction_run(
-        self, mocker: MockerFixture, checkpoint_result: CheckpointResult, mocked_posthog
-    ):
+    def test_run(self, mocker: MockerFixture, checkpoint_result: CheckpointResult, mocked_posthog):
         # Arrange
         context = mocker.Mock(spec=AbstractDataContext)
         set_context(context)
@@ -923,7 +790,7 @@ class TestV1ActionRun:
                     dry_run=False,
                     resource_identifiers=[
                         validation_identifier_a,
-                        ExpectationSuiteIdentifier(name=self.suite_a),
+                        ExpectationSuiteIdentifier(name=SUITE_A),
                     ],
                     site_names=site_names,
                 ),
@@ -932,7 +799,7 @@ class TestV1ActionRun:
                     dry_run=False,
                     resource_identifiers=[
                         validation_identifier_b,
-                        ExpectationSuiteIdentifier(name=self.suite_b),
+                        ExpectationSuiteIdentifier(name=SUITE_B),
                     ],
                     site_names=site_names,
                 ),
@@ -950,7 +817,7 @@ class TestV1ActionRun:
         }
 
     @pytest.mark.cloud
-    def test_UpdateDataDocsAction_run_cloud(
+    def test_run_with_cloud(
         self, mocker: MockerFixture, checkpoint_result: CheckpointResult, mocked_posthog
     ):
         # Arrange
@@ -993,7 +860,7 @@ class TestV1ActionRun:
                         validation_identifier_a,
                         GXCloudIdentifier(
                             resource_type=GXCloudRESTResource.EXPECTATION_SUITE,
-                            resource_name=self.suite_a,
+                            resource_name=self.SUITE_A,
                         ),
                     ],
                     site_names=site_names,
@@ -1005,7 +872,7 @@ class TestV1ActionRun:
                         validation_identifier_b,
                         GXCloudIdentifier(
                             resource_type=GXCloudRESTResource.EXPECTATION_SUITE,
-                            resource_name=self.suite_b,
+                            resource_name=self.SUITE_B,
                         ),
                     ],
                     site_names=site_names,
@@ -1016,101 +883,3 @@ class TestV1ActionRun:
             validation_identifier_a: {},
             validation_identifier_b: {},
         }
-
-    @pytest.mark.unit
-    def test_SlackNotificationAction_variable_substitution_webhook(
-        self, mock_context, checkpoint_result, mocked_posthog
-    ):
-        action = SlackNotificationAction(name="my_action", slack_webhook="${SLACK_WEBHOOK}")
-
-        with mock.patch("great_expectations.checkpoint.actions.send_slack_notification"):
-            action.run(checkpoint_result)
-
-        mock_context.config_provider.substitute_config.assert_called_once_with("${SLACK_WEBHOOK}")
-
-    @pytest.mark.unit
-    def test_SlackNotificationAction_variable_substitution_token_and_channel(
-        self, mock_context, checkpoint_result, mocked_posthog
-    ):
-        action = SlackNotificationAction(
-            name="my_action", slack_token="${SLACK_TOKEN}", slack_channel="${SLACK_CHANNEL}"
-        )
-
-        with mock.patch("great_expectations.checkpoint.actions.send_slack_notification"):
-            action.run(checkpoint_result)
-
-        assert mock_context.config_provider.substitute_config.call_count == 2
-        mock_context.config_provider.substitute_config.assert_any_call("${SLACK_CHANNEL}")
-        mock_context.config_provider.substitute_config.assert_any_call("${SLACK_TOKEN}")
-
-
-class TestActionIntegration:
-    """
-    Tests for actions that require external services to be available.
-    """
-
-    @pytest.fixture
-    def checkpoint_result(self, mocker) -> CheckpointResult:
-        return CheckpointResult(
-            run_id=RunIdentifier(run_name="e2e_test", run_time=dt.datetime.now(tz=dt.timezone.utc)),
-            run_results={
-                ValidationResultIdentifier(
-                    expectation_suite_identifier=ExpectationSuiteIdentifier(
-                        name="test-suite",
-                    ),
-                    run_id=RunIdentifier(run_name="test-run"),
-                    batch_identifier="test-datasource-test-asset",
-                ): ExpectationSuiteValidationResult(
-                    success=True,
-                    statistics={"successful_expectations": 3, "evaluated_expectations": 3},
-                    results=[],
-                    suite_name="test-suite",
-                ),
-                ValidationResultIdentifier(
-                    expectation_suite_identifier=ExpectationSuiteIdentifier(
-                        name="test-suite",
-                    ),
-                    run_id=RunIdentifier(run_name="test-run"),
-                    batch_identifier="test-datasource-test-asset",
-                ): ExpectationSuiteValidationResult(
-                    success=False,
-                    statistics={"successful_expectations": 2, "evaluated_expectations": 4},
-                    results=[],
-                    suite_name="test-suite",
-                ),
-            },
-            checkpoint_config=Checkpoint(
-                name="test-checkpoint",
-                validation_definitions=[
-                    mocker.MagicMock(spec=ValidationDefinition),
-                    mocker.MagicMock(spec=ValidationDefinition),
-                ],
-            ),
-            success=False,
-        )
-
-    @pytest.mark.integration
-    @pytest.mark.parametrize(
-        "action_cls, action_creds, expected_action_result",
-        [
-            pytest.param(
-                MicrosoftTeamsNotificationAction,
-                {"teams_webhook": "${GX_MS_TEAMS_WEBHOOK}"},
-                {"microsoft_teams_notification_result": "Microsoft Teams notification succeeded."},
-                id="ms_teams",
-            ),
-        ],
-    )
-    def test_actions(
-        self,
-        action_cls: type[ValidationAction],
-        action_creds: dict,
-        expected_action_result: dict,
-        checkpoint_result: CheckpointResult,
-    ):
-        # Necessary to retrieve config provider
-        gx.get_context(mode="ephemeral")
-
-        action = action_cls(name="test-action", **action_creds)
-        result = action.run(checkpoint_result=checkpoint_result)
-        assert result == expected_action_result
